Change driver behavior for missing logical unit.
This change removes an exception from the helper function _get_luid().
The exception occurs when the logical unit ID (LUID) is missing in
the sbdadm facility. Unfortunately there are scenarios where the
LUID will be missing and _get_luid() must complete for remove_export
to succeed.

If _get_luid() fails during create_export, the exception is still
thrown, with a slight change in the error message.

remove_export is now idempotent to the state of the LUID.

Change-Id: Iae472408c52a1cc28f23549b2028f820bb592c91