Fix SolarisISCSIDriver behavior for missing LUID
This change removes an exception from the SolarisISCSIDriver helper
function _get_luid(), which is supposed to fetch the logical unit
id (LUID) for an exported volume. The exception occurs when the
LUID is missing in the sbdadm facility. Unfortunately there are
scenarios where the LUID will be missing and _get_luid() must
complete for remove_export to succeed.

If _get_luid() fails during create_export, the exception is still
thrown, with a slight change in the error message.

remove_export is now idempotent to the state of the LUID.

Related-Bug: #1316352
Change-Id: Iae472408c52a1cc28f23549b2028f820bb592c91